#35: fixes main; fixes appveyor (abandon node < 6)
