+++ test/ignore.js
@@ -1,5 +1,5 @@
 // For old node.js versions, we use es5
-const ignore = require('../')
+const ignore = require('..')
 const {test} = require('tap')
 const cases = require('./fixtures/cases')
 
@@ -66,61 +66,3 @@ cases((
     t.end()
   })
 })
-
-test('.add(<Ignore>)', t => {
-  const a = ignore().add(['.abc/*', '!.abc/d/'])
-  const b = ignore().add(a).add('!.abc/e/')
-
-  const paths = [
-    '.abc/a.js',    // filtered out
-    '.abc/d/e.js',  // included
-    '.abc/e/e.js'   // included by b, filtered out by a
-  ]
-
-  t.deepEqual(a.filter(paths), ['.abc/d/e.js'])
-  t.deepEqual(b.filter(paths), ['.abc/d/e.js', '.abc/e/e.js'])
-  t.end()
-})
-
-test('fixes babel class', t => {
-  const {constructor} = ignore()
-
-  try {
-    constructor()
-  } catch (e) {
-    t.end()
-    return
-  }
-
-  t.is('there should be an error', 'no error found')
-  t.end()
-})
-
-test('#32', t => {
-  const KEY_IGNORE = typeof Symbol !== 'undefined'
-    ? Symbol.for('node-ignore')
-    : 'node-ignore'
-
-  const a = ignore().add(['.abc/*', '!.abc/d/'])
-
-  // aa is actually not an IgnoreBase instance
-  const aa = {}
-
-  /* eslint no-underscore-dangle: ["off"] */
-  aa._rules = a._rules.slice()
-  aa[KEY_IGNORE] = true
-
-  const b = ignore().add(aa).add('!.abc/e/')
-
-  const paths = [
-    '.abc/a.js',    // filtered out
-    '.abc/d/e.js',  // included
-    '.abc/e/e.js'   // included by b, filtered out by a
-  ]
-
-  t.deepEqual(a.filter(paths), ['.abc/d/e.js'])
-  t.deepEqual(b.filter(paths), ['.abc/d/e.js', '.abc/e/e.js'])
-  t.end()
-})
-
-test('')
