+++ test/others.js
@@ -1,0 +1,60 @@
+const ignore = require('..')
+const {test} = require('tap')
+
+test('.add(<Ignore>)', t => {
+  const a = ignore().add(['.abc/*', '!.abc/d/'])
+  const b = ignore().add(a).add('!.abc/e/')
+
+  const paths = [
+    '.abc/a.js',    // filtered out
+    '.abc/d/e.js',  // included
+    '.abc/e/e.js'   // included by b, filtered out by a
+  ]
+
+  t.deepEqual(a.filter(paths), ['.abc/d/e.js'])
+  t.deepEqual(b.filter(paths), ['.abc/d/e.js', '.abc/e/e.js'])
+  t.end()
+})
+
+test('fixes babel class', t => {
+  const {constructor} = ignore()
+
+  try {
+    constructor()
+  } catch (e) {
+    t.end()
+    return
+  }
+
+  t.is('there should be an error', 'no error found')
+  t.end()
+})
+
+test('#32', t => {
+  const KEY_IGNORE = typeof Symbol !== 'undefined'
+    ? Symbol.for('node-ignore')
+    : 'node-ignore'
+
+  const a = ignore().add(['.abc/*', '!.abc/d/'])
+
+  // aa is actually not an IgnoreBase instance
+  const aa = {}
+
+  /* eslint no-underscore-dangle: ["off"] */
+  aa._rules = a._rules.slice()
+  aa[KEY_IGNORE] = true
+
+  const b = ignore().add(aa).add('!.abc/e/')
+
+  const paths = [
+    '.abc/a.js',    // filtered out
+    '.abc/d/e.js',  // included
+    '.abc/e/e.js'   // included by b, filtered out by a
+  ]
+
+  t.deepEqual(a.filter(paths), ['.abc/d/e.js'])
+  t.deepEqual(b.filter(paths), ['.abc/d/e.js', '.abc/e/e.js'])
+  t.end()
+})
+
+test('')
